FIX: move zeitwerk config to main lib
